+++ metadata.rb
@@ -7,7 +7,7 @@ maintainer_email 'jonathan.hartman@socrata.com'
 license          'apache2'
 description      'Configures a node to be a bastion host'
 long_description 'Configures a node to be a bastion host'
-version          '0.3.1'
+version          '1.0.0'
 
 supports         'ubuntu'